Refactor:backup operation by test db connection inside the DatabaseBackupService, and move testConnection method to TestDatabaseConnectionService instead of being in DatabaseConnectionControllers
