+++ routes/api.php
@@ -4,10 +4,8 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\DatabaseConnectionController;
 use App\Http\Controllers\BackupJobController;
+use App\Services\TestDatabaseConnectionService;
 
-Route::get('database-connections/test/{db_id}',[DatabaseConnectionController::class, 'testConnection']);
 Route::apiResource('database-connections',DatabaseConnectionController::class);
 Route::apiResource('backup-jobs',BackupJobController::class)->except(['update']);
-
-
-
+Route::get('test-db-connection/{id}',[TestDatabaseConnectionService::class,'testConnection']);
